Couple more items in demonstrated skills
Added mention of @property decorator as I'm adding that in on the WPHomepage class. Also mentioning the WPHomepage class as I'm using that to dictate overall structure of the page, which lets me focus test cases down some more. (This is a practice we used at Big Fish in the Python framework I worked on there.)